# MvcAspAzure/swagger.yaml
openapi: 3.0.0
info:
  title: Warehouse & Shipment API
  version: 1.0.0

paths:
  /shipments:
    get:
      summary: Get all shipments
      operationId: getAllShipments
      responses:
        '200':
          description: A list of shipments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shipment'

    post:
      summary: Create a new shipment
      operationId: createShipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShipmentCreate'
      responses:
        '201':
          description: Shipment created

  /shipments/{id}:
    get:
      summary: Get shipment by ID
      operationId: getShipmentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shipment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'

    delete:
      summary: Delete shipment
      operationId: deleteShipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Shipment deleted

  /warehouses:
    get:
      summary: Get all warehouses
      operationId: getAllWarehouses
      responses:
        '200':
          description: A list of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Warehouse'

    post:
      summary: Create a new warehouse
      operationId: createWarehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseCreate'
      responses:
        '201':
          description: Warehouse created

  /warehouses/{id}:
    get:
      summary: Get warehouse by ID
      operationId: getWarehouseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Warehouse found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'

    delete:
      summary: Delete warehouse
      operationId: deleteWarehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Warehouse deleted


# /warehouses/{id}/activate:
#   post:
#     summary: Activate a warehouse
#     operationId: activateWarehouse
#     parameters:
#       - name: id
#         in: path
#         required: true
#         schema:
#           type: integer
#     responses:
#       '204':
#         description: Warehouse activated

# /warehouses/search:
#   get:
#     summary: Search warehouses by filters
#     operationId: searchWarehouses
#     parameters:
#       - name: placeId
#         in: query
#         schema:
#           type: integer
#       - name: name
#         in: query
#         schema:
#           type: string
#     responses:
#       '200':
#         description: Filtered warehouse list
#         content:
#           application/json:
#             schema:
#               type: array
#               items:
#                 $ref: '#/components/schemas/Warehouse'



components:
  schemas:
    Shipment:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string

    ShipmentCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        status:
          type: string

    Warehouse:
      type: object
      properties:
        id:
          type: string
        location:
          type: string

    WarehouseCreate:
      type: object
      required:
        - location
      properties:
        location:
          type: string